+++ src/renderer/pages/settings/advanced_app_settings/build_info.tsx
