+++ src/renderer/pages/settings/advanced_app_settings/network_diagnostics/cgnat_command_section.tsx
