+++ src/renderer/pages/settings/advanced_app_settings/network_diagnostics/nat_type_section.tsx
