+++ src/renderer/pages/settings/create.tsx
@@ -1,14 +1,12 @@
-import { DolphinLaunchType } from "@dolphin/types";
-
-import { AdvancedAppSettings } from "@/containers/Settings/AdvancedAppSettings";
-import { ChatOptions } from "@/containers/Settings/ChatOptions";
-import { DolphinSettings } from "@/containers/Settings/DolphinSettings";
-import { HelpPage } from "@/containers/Settings/HelpPage";
-import { MeleeOptions } from "@/containers/Settings/MeleeOptions";
-import { ReplayOptions } from "@/containers/Settings/ReplayOptions";
-import type { SettingSection } from "@/containers/Settings/types";
-
+import { AdvancedAppSettings } from "./advanced_app_settings/advanced_app_settings";
+import { ChatSettings } from "./chat_settings/chat_settings";
+import { NetplayDolphinSettings } from "./dolphin_settings/netplay_dolphin_settings";
+import { PlaybackDolphinSettings } from "./dolphin_settings/playback_dolphin_settings";
+import { GameSettings } from "./game_settings/game_settings";
+import { HelpPage } from "./help_page/help_page";
+import { ReplaySettings } from "./replay_settings/replay_settings";
 import { SettingsPage } from "./settings_page";
+import type { SettingSection } from "./types";
 
 export function createSettingsPage(): { Page: React.ComponentType } {
   const settings: SettingSection[] = [
@@ -18,17 +16,17 @@ export function createSettingsPage(): { Page: React.ComponentType } {
         {
           name: "Game",
           path: "melee-options",
-          component: <MeleeOptions />,
+          component: <GameSettings />,
         },
         {
           name: "Replays",
           path: "replay-options",
-          component: <ReplayOptions />,
+          component: <ReplaySettings />,
         },
         {
           name: "Chat",
           path: "chat-options",
-          component: <ChatOptions />,
+          component: <ChatSettings />,
         },
       ],
     },
@@ -38,12 +36,12 @@ export function createSettingsPage(): { Page: React.ComponentType } {
         {
           name: "Netplay",
           path: "netplay-dolphin-settings",
-          component: <DolphinSettings dolphinType={DolphinLaunchType.NETPLAY} />,
+          component: <NetplayDolphinSettings />,
         },
         {
           name: "Playback",
           path: "playback-dolphin-settings",
-          component: <DolphinSettings dolphinType={DolphinLaunchType.PLAYBACK} />,
+          component: <PlaybackDolphinSettings />,
         },
       ],
     },
@@ -68,5 +66,7 @@ export function createSettingsPage(): { Page: React.ComponentType } {
     },
   ];
 
-  return { Page: () => <SettingsPage settings={settings} /> };
+  const Page = () => <SettingsPage settings={settings} />;
+
+  return { Page };
 }
