+++ src/renderer/pages/settings/dolphin_settings/gecko_codes/add_codes/add_codes.container.tsx
@@ -1,7 +1,7 @@
 import type { GeckoCode } from "@dolphin/config/geckoCode";
 import { parseGeckoCodes } from "@dolphin/config/geckoCode";
 
-import { AddCodes } from "./AddCodes";
+import { AddCodes } from "./add_codes";
 
 export const AddCodesContainer = ({
   existingGeckoCodeNames = [],
