+++ src/renderer/pages/settings/dolphin_settings/gecko_codes/add_codes/add_codes.stories.tsx
@@ -1,6 +1,6 @@
 import type { ComponentMeta, ComponentStory } from "@storybook/react";
 
-import { AddCodes } from "./AddCodes";
+import { AddCodes } from "./add_codes";
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
