+++ src/renderer/pages/settings/dolphin_settings/gecko_codes/add_codes/add_codes.tsx
