+++ src/renderer/pages/settings/dolphin_settings/gecko_codes/manage_codes/manage_codes.tsx
