+++ src/renderer/pages/settings/dolphin_settings/gecko_codes/tabbed_dialog.tsx
