+++ src/renderer/pages/settings/dolphin_settings/netplay_dolphin_settings.tsx
@@ -1,0 +1,7 @@
+import { DolphinLaunchType } from "@dolphin/types";
+
+import { DolphinSettings } from "./dolphin_settings";
+
+export const NetplayDolphinSettings = () => {
+  return <DolphinSettings dolphinType={DolphinLaunchType.NETPLAY} />;
+};
